# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "应用程序的简短描述.",
        "title": "Swagger Gin-Basic-Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/custom/routing/table_manage/create": {
            "post": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "创建新的表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "table_manage"
                ],
                "summary": "创建表",
                "parameters": [
                    {
                        "description": "表信息",
                        "name": "table",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TableManageAdd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/custom/routing/table_manage/query": {
            "post": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "获取表的信息列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "table_manage"
                ],
                "summary": "获取表信息列表",
                "parameters": [
                    {
                        "description": "表信息",
                        "name": "table",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.PageInfo"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "query_info": {
                                            "$ref": "#/definitions/model.TableManageQuery"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/custom/routing/table_manage/{id}": {
            "get": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "获取指定表的信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "table_manage"
                ],
                "summary": "获取表信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "表ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/custom/routing/table_manage/{id}/delete": {
            "delete": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "删除指定表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "table_manage"
                ],
                "summary": "删除表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "表ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/custom/routing/table_manage/{id}/update": {
            "put": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "更新指定表的信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "table_manage"
                ],
                "summary": "更新表信息",
                "parameters": [
                    {
                        "description": "表信息",
                        "name": "table",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TableManageUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "登录接口",
                "parameters": [
                    {
                        "description": "表信息",
                        "name": "table",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/global.Operator"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "global.Operator": {
            "type": "object",
            "properties": {
                "c_user_name": {
                    "description": "姓名",
                    "type": "string"
                },
                "c_user_number": {
                    "description": "工号",
                    "type": "string"
                }
            }
        },
        "model.FieldManageAdd": {
            "type": "object",
            "properties": {
                "b_enable": {
                    "description": "是否可用",
                    "type": "integer"
                },
                "b_foreign_key": {
                    "description": "是否为外键 0：非外键 1：外键",
                    "type": "integer"
                },
                "b_primary_key": {
                    "description": "是否为主键 0：非主键 1：主键",
                    "type": "integer"
                },
                "b_synchronous": {
                    "description": "是否已同步 0：未同步 1：已同步",
                    "type": "integer"
                },
                "c_association_field": {
                    "description": "关联字段",
                    "type": "integer"
                },
                "c_association_table": {
                    "description": "关联表",
                    "type": "integer"
                },
                "c_default_value": {
                    "description": "默认值",
                    "type": "string"
                },
                "c_field": {
                    "description": "列名",
                    "type": "string"
                },
                "c_field_details": {
                    "description": "列说明",
                    "type": "string"
                },
                "c_field_name": {
                    "description": "列名称",
                    "type": "string"
                },
                "c_field_type": {
                    "description": "列类型",
                    "type": "string"
                },
                "c_primary_rule": {
                    "description": "主键生成规则 input：手动输入 auto_increment：自增长 snowflake_id：雪花ID",
                    "type": "string"
                },
                "c_table_id": {
                    "description": "表id，关联表：table_manage.id",
                    "type": "integer"
                },
                "i_field_length": {
                    "description": "字段长度",
                    "type": "integer"
                }
            }
        },
        "model.FieldManageUpdate": {
            "type": "object",
            "properties": {
                "b_enable": {
                    "description": "是否可用",
                    "type": "integer"
                },
                "b_foreign_key": {
                    "description": "是否为外键 0：非外键 1：外键",
                    "type": "integer"
                },
                "b_primary_key": {
                    "description": "是否为主键 0：非主键 1：主键",
                    "type": "integer"
                },
                "b_synchronous": {
                    "description": "是否已同步 0：未同步 1：已同步",
                    "type": "integer"
                },
                "c_association_field": {
                    "description": "关联字段",
                    "type": "integer"
                },
                "c_association_table": {
                    "description": "关联表",
                    "type": "integer"
                },
                "c_default_value": {
                    "description": "默认值",
                    "type": "string"
                },
                "c_field": {
                    "description": "列名",
                    "type": "string"
                },
                "c_field_details": {
                    "description": "列说明",
                    "type": "string"
                },
                "c_field_name": {
                    "description": "列名称",
                    "type": "string"
                },
                "c_field_type": {
                    "description": "列类型",
                    "type": "string"
                },
                "c_primary_rule": {
                    "description": "主键生成规则 input：手动输入 auto_increment：自增长 snowflake_id：雪花ID",
                    "type": "string"
                },
                "c_table_id": {
                    "description": "表id，关联表：table_manage.id",
                    "type": "integer"
                },
                "i_field_length": {
                    "description": "字段长度",
                    "type": "integer"
                },
                "id": {
                    "description": "编号",
                    "type": "integer"
                }
            }
        },
        "model.TableManageAdd": {
            "type": "object",
            "properties": {
                "b_enable": {
                    "description": "是否可用，0：不可用，1：可用",
                    "type": "integer"
                },
                "c_table": {
                    "description": "表名",
                    "type": "string"
                },
                "c_table_details": {
                    "description": "表说明",
                    "type": "string"
                },
                "c_table_name": {
                    "description": "表名称",
                    "type": "string"
                },
                "fields": {
                    "description": "列信息",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FieldManageAdd"
                    }
                }
            }
        },
        "model.TableManageQuery": {
            "type": "object",
            "properties": {
                "c_table": {
                    "description": "表名",
                    "type": "string"
                },
                "c_table_name": {
                    "description": "表名称",
                    "type": "string"
                },
                "id": {
                    "description": "编号",
                    "type": "integer"
                }
            }
        },
        "model.TableManageUpdate": {
            "type": "object",
            "properties": {
                "b_enable": {
                    "description": "是否可用，0：不可用，1：可用",
                    "type": "integer"
                },
                "c_table": {
                    "description": "表名",
                    "type": "string"
                },
                "c_table_details": {
                    "description": "表说明",
                    "type": "string"
                },
                "c_table_name": {
                    "description": "表名称",
                    "type": "string"
                },
                "fields": {
                    "description": "列信息",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FieldManageUpdate"
                    }
                },
                "id": {
                    "description": "编号",
                    "type": "integer"
                }
            }
        },
        "util.PageInfo": {
            "type": "object",
            "properties": {
                "page": {
                    "description": "当前页码",
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "description": "每页记录数",
                    "type": "integer",
                    "example": 10
                },
                "query_info": {
                    "description": "自定义查询条件"
                },
                "total_count": {
                    "description": "总记录数",
                    "type": "integer",
                    "example": 0
                }
            }
        }
    },
    "securityDefinitions": {
        "token": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}